requirements.txt
pyramid
pyramid_mailer
pyramid_signup
pyramid_deform
pyramid_beaker
pyramid_jinja2
pyramid_tm
pyramid_debugtoolbar
pyramid_webassets
pyramid_socketio
simplejson
redis

zope.sqlalchemy
SQLAlchemy
transaction
gunicorn
pycco

deform
webassets
pystache
colander
pytz
cryptacular
deform_jinaj2
gevent-socketio
gevent-websocket
